requirements.txt
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
uvloop==0.14.0
PyYAML==5.3
asyncpgsa==0.26.1
webargs==5.5.3
pytest-aiohttp==0.3.0
psycopg2-binary==2.8.4